v1.2.1 (#22)
* [[explicit-variables]] Adds command to toggle explicit variable checks (#17)

This patch adds a setting to allow the user to toggle variable checking 
for the linter to alert the user if variables have not been declared.

* [[indentation]] Adds support for custom indentaton (#15)

This patch adds support for custom indentation that allows you to choose 
between tabs or spaces, with custom indentation size for spaces. Tabs are 
locked at 1 indentation size.

Co-authored-by: William <46687933+WilliamDelRey@users.noreply.github.com>

* [[live-server]] Edited code in vscode is displayed in livecode scripts (#16)

This patch adds the connection between vscode and livecode using TCP 
sockets to send code changes from vscode to the LiveCode script editor.

Co-authored-by: William <46687933+WilliamDelRey@users.noreply.github.com>

* Version updated

* [[live-server]] Adds toast notifications for users (#19)

This patch adds toast notifications to indicate to the user what is 
happening in the connection between vscode and LiveCode. Also tidied up 
some functions.

* [[indentation]] Changed indentation to be same format as current code (#20)

This patch adds the indentation checks to be within the switch statement 
of the already existing code.

* Version updated

* fix:missing comma

* fix: missing } due to conflict

Co-authored-by: Ruaridh Bell <107618943+livecoderuaridh@users.noreply.github.com>